updated Xpath and moved kmLocators.js from utils to locatorPackage
